chore(rpc): move binary_search into pub method
